CBMC proofs now run on Linux and Windows.
The process for building and running the proofs was designed for Linux.
Customers and developers work primarily on Windows.  This change allows
customers and developers to run (and write) proofs on their own in their
own build environment.  The method uses python scripts and templates to
construct makefiles for either Linux make or Visual Studio nmake.