Fix CBMC ProcessDHCPReplies proof by updating cbmc-batch.yaml correctly.
The CBMC proofs run on AWS Batch are driven by a configuration file
cbmc-batch.yaml.  That file needs to be updated when the Makefiles are
updated, and needs to be updated to include the --unwinding-assertions
flag in the list of CBMCFLAGS.  This commit fixes the cbmc-batch.yaml
generation and generates the correct cbmc-batch.yaml file.