Change CBMC top-level proof paths to r1.5 layout
This commit forms the first part of a patchlist to change the paths used
in CBMC proofs so that they conform to the new directory layout
introduced in a:FR release-1.5.

This commit changes the paths in the top-level Makefiles and the patch
files that get applied to the a:FR codebase before CI is run.