Add WiFi Security value check in quick start script (#822)
The quick start script uses configuration values supplied by the user in
configure.json file. The WiFi security value is not a free form string
and must be one of the following values: eWiFiSecurityOpen,
eWiFiSecurityWEP, eWiFiSecurityWPA, eWiFiSecurityWPA2. This change makes
sure that we terminate the script and provide a useful error message if
the user has supplied incorrect WiFi security value.

Signed-off-by: Gaurav Aggarwal <aggarg@amazon.com>