Add option to run uncrustify to the pre-commit script
pre-commit script runs some code style checks including uncrustify
before a commit is made. This commit adds 2 options to the script to
allow users to use the same script to uncrustify files:

1. python tools/git/hooks/src/pre_commit.py --uncrustify
2. python tools/git/hooks/src/pre_commit.py --uncrustify-all-files

The first one uncrustifies the modified files by the user while the
second one uncrustifies all the files in the repo.

Signed-off-by: Gaurav Aggarwal <aggarg@amazon.com>