Remove call to vTaskDelete from runDemoTask
runDemoTask is created using Iot_CreateDetachedThread API and the
threads created using this API must not call vTaskDelete as it is
automatically called when the thread routine returns. Explicitly calling
vTaskDelete results in a memory leak as the internally allocated
threadInfo_t does not get freed.

This was reported here:
https://github.com/aws/amazon-freertos/issues/989

Signed-off-by: Gaurav Aggarwal <aggarg@amazon.com>