Limit Tx/Rx task pairs to 1 for MQTT_Agent_MultiTaskTest on Mediatek
The default number of Tx/Rx task pairs for MQTT_Agent_MultiTaskTest is 2
which means that we try to establish two simultaneous connections to the
AWS IoT broker. This results in test failure as we do not have enough
RAM to do two simultaneous TLS handshakes. This change limits the number
of Tx/Rx task pairs to 1 for Mediatek.

This change also sets configUSE_MALLOC_FAILED_HOOK to 1 which makes memory
failure debugging a lot easier.

Signed-off-by: Gaurav Aggarwal <aggarg@amazon.com>