Fix Greengrass test and add debugging support
Each of the Greengrass test calls IotSdk_Init in the setup and
IotSdk_Cleanup in the teardown. IotSdk_Init creates the system task pool
which consists of 2 tasks of 2K stack size each. IotSdk_Cleanup deletes
those tasks by calling vTaskDelete. vTaskDelete only marks a task for
deletion and the actual deletion happens in the idle task. The test
runner task for Mediatek was running at tskIDLE_PRIORITY + 1 and was not
allowing the Idle task to run between test runs. As a result, the heap
space allocated for system pool tasks' stack and TCB was not getting
freed. This was resulting in memory allocation failures leading to test
failures. This change lowers the priority of the test runner task to
tskIDLE_PRIORITY to ensure that the heap space is freed between test
runs. This fixes the Greengrass tests for Mediatek.

The debugging instruction on the getting started page for Mediatek
( https://docs.aws.amazon.com/freertos/latest/userguide/getting_started_mediatek.html )
mention MTK_DEBUGGER macro and wait_ice variable. These are needed for
debugging this board but were missing from the code. This change adds
these macros and variables to enable debugging suopport.

Signed-off-by: Gaurav Aggarwal <aggarg@amazon.com>