Add demo start and end markers
The following messages are printed before starting and finishing the
demos:

* ---------STARTING DEMO---------
* -------DEMO FINISHED-------

These messages are used by the test framework to determine when a demo
starts and when it finishes. If these messages are not available, the
test framework has to rely on a timeout which may not be accurate
always.

Signed-off-by: Gaurav Aggarwal <aggarg@amazon.com>