Refactor prvCheckoptions on master branch. (#1481)
* Refactor the function prvCheckOptions.

Refactor the function prvCheckOptions so that the two bodies of the
two nested loops as functions replacing the loop bodies.  This makes
the code easier to read.  It also makes it possible to prove the
memory safety of the code with CBMC quickly.

* Add CBMC patch removing static attributes from prvCheckOptions functions for CBMC proof.

Also remove CBMC patch refactoring prvCheckOptions (this branch does the refactoring).