Use CMake for building & running CBMC proofs (#1537)
This commit introduces:

- a CBMC 'toolchain' (needed for compiling CBMC proofs);
- a file containing common 'recipes' for building and running each
  proof.

* Break goto-instrument commands into separate steps

We previously applied a series of goto-instrument transformations to a
goto binary in one single CMake command. Splitting the command up allows
us to print more informative status messages, interleaved with any
output that goto-instrument might print.

* Don't use bogus loop for CBMC unwindset

The --unwindset flag to CBMC requires an argument, but many proofs don't
need an unwindset. This commit ensures that --unwindset is not used at
all if a proof's unwindset is empty, instead of using a bogus loop ID as
an argument. This makes it easier to see if a real loop ID was
misspelled if/when CBMC adds such a feature.

* indentation

* Allow proofs to use --nondet-static{,-except}

Proof-writers may now set either the `cbmc_nondet_static` variable to ON
(or OFF, which is the default), or add variables to the
`cbmc_nondet_static_exclude` list. This will insert a goto-instrument
pass that makes all static variables in the program start with an
unconstrained value, with the exception of any variable in the
`cbmc_nondet_static_exclude` list.

This has the effect that CBMC will not assume that static variables have
been initialized (e.g. to zero), but will check the program for all
possible initial values of static variables.

* Replace 'all-proofs' target with 'cbmc'

* Addressed markrtuttle review comments

* Use platform-agnostic invocations of CBMC

* Backported IotHttpsClient_SendSync fix from master

* Add ability to export static functions

This is used when writing harnesses for static functions. Normally, it
would not be possible to link the harness to the file containing to the
static functions; setting cbmc_export_static to TRUE in a particular
proof ensures that proof-writers can link against static functions.