Merge release to master (#1773)
* Bringing changes for AFR Release 202002.00 (#1723)

* Fix a dependency parsing bug in CMake from #1161 (#1655)

In previous PR when we add a feature to disable building demos and
test, we add a check to not search dependencies of aws_demos and
aws_tests target when demos or tests are disabled. But we forgot to also
move adding the dependencies to demos_base and tests_base inside the
check. As a result, the "dependencies" from previous for loop was added
to demos_base and tests_base.

* Add board dir for cypress in CMake (#1684)

* Merge pull request #1700 from tgsong/ota_e2e_s3_cleanup

Only delete files uploaded in OTA E2E

* OTA e2e scripts: Replace dots with underscores (#1717)

* Update FreeRTOS-Kernel submodule branch to release-candidate

Update FreeRTOS-Kernel submodule branch from V10.2.1-convergence-FreeRTOS-Source
to release-candidate_10.3.0_rc3
commit id: d7550ef1a0560ceddac7c0b8bc0148272ac641a6

Co-authored-by: Tiangang Song <ts.whu@outlook.com>
Co-authored-by: Pavan Madehalli Ranganath <rangpava@amazon.com>

* Variable name change from uxPendedTicks to xPendedTicks. (#1724)

Due to the kernel version bump, Tracealyzer also needs to be updated. This is to fix the build failure on Windows and Microchip in this PR -- https://github.com/aws/amazon-freertos/pull/1723

* Update FreeRTOS-Kernel submodule to V10.3.0-kernel-only (#1743)

Update FreeRTOS-Kernel submodule branch from release-candidate_10.3.0_rc3 to V10.3.0-kernel-only
commit id: 210b1ffcc87bcff93871a37fbf0ad2033870ecaf

* Change http echo server from 'httpbin.org' to 'postman-echo.com' (#1742)

HTTPS_Client_System test Group is fails for TI when connecting to
'httpbin.org'. Connect to 'postman-echo.com' to solve this issue.

* Update checksums.json file (#1744)

* Update CHANGELOG.md (#1748)

* Change name from "Amazon FreeRTOS" to "FreeRTOS" (#1736)

* Cypress kernel dependency workaround (#1750)

* From FreeRTOS kernel 10.2.0, copying over Wiced_CY port to vendor directory.

* Make Cortex R4 target use the kernel files at new location.

* The project file update probably won't matter, but in IDE at least you'll now get the correct source files.

* Change Nordic kernel portable files path (#1751)

* Change paths for Nordic test and demo projects to build with kernel

kernel_branch: release-candidate_10.3.0_rc3
commit id: d7550ef1a0560ceddac7c0b8bc0148272ac641a6

* Change Cmake and IDE to use kernel portable files from Nordic SDK

* Temporary fix for Renesas test/demo failure

* Temporary fix for Renesas test/demo failure

* Changes for 202002.00 Release (#1756)

* Fix: GGD Test failures (#1616)

* Fix: GGD Test failures on ST and other boards

GGD_JSONRequestGetFile() would not receive the entire file
in one call on some boards. This adds a loop to receive the
file in chunks if the request was not completed.

* Add logging statement to GGD File Retrieval

* Unapologetically deleting WiFi test cases. (#1685)

* Unapologetically deleting WiFi test cases that don't do much. Test case references are left in commented out. Consider this shame mark.

* uncrustify.

* Deleting all references to unused references.

* Update Major number for Release from 201912 to 202002

* Update checksum file for 202002.00 release

Co-authored-by: Muneeb Ahmed <54290492+muneebahmed10@users.noreply.github.com>
Co-authored-by: Yuhui.Zheng <10982575+yuhui-zheng@users.noreply.github.com>

* Change uxPendedTicks to xPendedTicks in all proofs (#1758)

This ensures that CBMC proofs are able to build after the name-change in
commit 23839bb.

* ECC key import fix. (#1760)

* Add FreeRTOS_Porting_Guide.pdf and FreeRTOS_Qualification_Guide.pdf (#1764)

* Change release date from 2/17/2020 to 2/18/2020 (#1769)

Co-authored-by: Tiangang Song <ts.whu@outlook.com>
Co-authored-by: Pavan Madehalli Ranganath <rangpava@amazon.com>
Co-authored-by: Yuhui.Zheng <10982575+yuhui-zheng@users.noreply.github.com>
Co-authored-by: yngki <47755894+yngki@users.noreply.github.com>
Co-authored-by: Ravishankar Bhagavandas <bhagavar@amazon.com>
Co-authored-by: AniruddhaKanhere <60444055+AniruddhaKanhere@users.noreply.github.com>
Co-authored-by: Muneeb Ahmed <54290492+muneebahmed10@users.noreply.github.com>
Co-authored-by: Kareem Khazem <karkhaz@karkhaz.com>
Co-authored-by: lundinc2 <53273776+lundinc2@users.noreply.github.com>