Removed Unused Struct Members Related to Find_Objects. (#1809)
These two struct members, xFindObjectInit, and xFindObjectComplete, are undocumented and unused. There is no reference in the code to xFindObjectInit being used.
xFindObjectComplete is actually used, but the code path it is used in is either impossible, or a bug. Since the P11SessionPtr_t struct is memset to 0
immediately after being malloc'd the line of code using xFindObjectComplete will always evaluate to false, as it is comparing it to CK_TRUE,
as CK_TRUE is equal to 1, and xFindObjectComplete is never explicitly set to 1.