Fix demos on Marvell (#2633)
Problem: Many demos were failing in Marvell board due to lack of memory in heap while establishing a TLS connection.
Root Cause: Establishing a TLS connection requires a good amount of memory in the HEAP. The tasks created with xTaskCreate() function also allocates the memory for the stack in the heap. Marvell's heap(~73KB) was running out of space due to these allocations.
Fix: Optimized the stack sizes for different tasks for the demo and verified the demos.