Update logging to not print metadata in line breaks for WinSim ports (#2760)
A newline was printed with metadata because it was configured to only clear the message buffer for LF rather than CRLF. In addition, the message number and tick count were always printed even though the log was on the same line. This is updated so that information is only logged with each CRLF.

This follows changes from FreeRTOS/FreeRTOS#397.