Fix indefinite wait issue in IotNetworkAfr_Destroy() (#3029)
As part of the change in PR #2986, the `IotNetworkAfr_Close` function was updated to wait for the receive task to be exited and clear the `_FLAG_RECEIVE_TASK_EXITED` flag as part of the `xEventGroupWaitBits()` call. However, that affects the `IotNetworkAfr_Destroy` function to wait indefinitely as it also contains logic to block for the receive task to exit. 

As the `IotNetworkAfr_Destroy` function is supposed to be called ONLY _after_ the connection is closed (as mentioned [here](https://github.com/aws/amazon-freertos/blob/201912.00/libraries/abstractions/platform/include/platform/iot_network.h\#L269-L270)), the receive task should have already exited (because of call to `IotNetworkAfr_Close`) before `IotNetworkAfr_Destroy` function is called, and thus, the wait for receive task to be exited is not required.