common-io UART fixes for Espressif's SoCs (#3435)
* esp boards: Enable uart flow control only if IDF console is not
initialised on UART0

Add iot_uart_init() in aws_tests application for ESP32-S2

* espressif/common_io: Save default configuration in context after
successfully initialising UART

Update UART pin map with pins common to ESP32 and ESP32-S2

* esp common_io uart: initialize stack variable

Zero uart_config variable on the stack so that any fields not explicitly
set will be zero rather than undefined.

Signed-off-by: Shubham Kulkarni <shubham.kulkarni@espressif.com>

* esp common_io uart: fix word length handling

The ioctl for configuring the UART was setting the data_bits member of
uart_config_t directly from the value in an IotUARTConfig_t ucWordlength
member. This is incorrect because the former is an enum and the latter
is an unsigned integer. Code for performing the translation has been
added.

Signed-off-by: Shubham Kulkarni <shubham.kulkarni@espressif.com>

* esp common_io uart: fix bug in flow control config

When flow control was requested, the uart_config_t member
rx_flow_ctrl_thresh was being set to UART_FIFO_LEN. This was leading to
failures in uart_param_config() in esp-idf/components/driver/uart.c
because there is a check to ensure that uart_config->rx_flow_ctrl_thresh
< UART_FIFO_LEN.

For an explanation of why the threshold must be less than the FIFO
length, see the esp32_technical_reference_manual_en.pdf document which
in section 13.3.6.1 Hardware Flow Control says:

  If data in Rx_FIFO is greater than UART_RX_FLOW_THRHD, the level of
  rtsn_out will be lowered.

The amount of data in the FIFO cannot be greater than the length of the
FIFO, so the maximum value that the threshold can be set to is one less
than the length of the FIFO.

Signed-off-by: Shubham Kulkarni <shubham.kulkarni@espressif.com>

* esp common_io uart: fix error code handling

In the handling for the eUartSetConfig ioctl, if uart_driver_delete or
iot_uart_driver_install failed, but uart_register_callback_with_isr
passed, it was possible for the ioctl operation to appear successful
even though it had failed.

Signed-off-by: Shubham Kulkarni <shubham.kulkarni@espressif.com>

* esp common_io uart: only cache config on success

When configuring the uart using the eUartSetConfig ioctl, the cached
value of the UART configuration should only be written if the
configuration is successfully applied.

Signed-off-by: Shubham Kulkarni <shubham.kulkarni@espressif.com>

Co-authored-by: David Frey <dfrey@sierrawireless.com>